Refactor process builders on Borrow Log
The logic contained in the Update Book Copy Status was partitioned into
two new processes: Create Borrow Log & Update Borrow Log. This was done
to future-proof processes on Borrow Log and to provide a guaranteed
order of execution. Proccesses that are triggered on a Create action
should defined in the Create Borrow Log process. Likewise, processes
triggered on update should be defiend in the Update Borrow Log process.

Added a new validation rule on Borrow Log, newRecordReturnedStatusFalse.
This validation rule prevents users from creating a new Borrow Log
record with the Returned field set to true.

Cleaned up the formatting of validateBookCopyAvailability and
validateReturnedField formulas.