Change Book Copy Status field to read-only
The Book Copy Status field was made read-only to prevent users from manually
changing the availability of a book copy. This was achieved by creating a
formula field in Book Copy that references the Status field. The Status field
was removed from the Book Copy page layout, and the formula field was added to
the layout. Read access to the formula field was added to the Cheltham Admin
Access permission set.